# apps/growup-backend/src/generated/swagger.json
{
	"components": {
		"examples": {},
		"headers": {},
		"parameters": {},
		"requestBodies": {},
		"responses": {},
		"schemas": {
			"IPlant": {
				"properties": {
					"_id": {
						"properties": {
							"$oid": {
								"type": "string"
							}
						},
						"required": [
							"$oid"
						],
						"type": "object"
					},
					"id": {
						"type": "string"
					},
					"name": {
						"type": "string"
					},
					"description": {
						"type": "string"
					},
					"type": {
						"type": "string"
					},
					"sunlight": {
						"type": "string"
					},
					"wateringFrequency": {
						"type": "string"
					},
					"wateringDescription": {
						"type": "string"
					},
					"imageUri": {
						"type": "string"
					},
					"wateringDates": {
						"items": {
							"type": "string"
						},
						"type": "array"
					}
				},
				"required": [
					"_id",
					"id",
					"name",
					"description",
					"type",
					"sunlight",
					"wateringFrequency",
					"wateringDescription",
					"imageUri",
					"wateringDates"
				],
				"type": "object",
				"additionalProperties": false
			},
			"NotFoundError": {
				"properties": {
					"name": {
						"type": "string"
					},
					"message": {
						"type": "string"
					},
					"stack": {
						"type": "string"
					}
				},
				"required": [
					"name",
					"message"
				],
				"type": "object",
				"additionalProperties": false
			},
			"UnauthorizedError": {
				"properties": {
					"name": {
						"type": "string"
					},
					"message": {
						"type": "string"
					},
					"stack": {
						"type": "string"
					}
				},
				"required": [
					"name",
					"message"
				],
				"type": "object",
				"additionalProperties": false
			},
			"Pick_IPlant.Exclude_keyofIPlant._id__": {
				"properties": {
					"id": {
						"type": "string"
					},
					"name": {
						"type": "string"
					},
					"description": {
						"type": "string"
					},
					"type": {
						"type": "string"
					},
					"sunlight": {
						"type": "string"
					},
					"wateringFrequency": {
						"type": "string"
					},
					"wateringDescription": {
						"type": "string"
					},
					"imageUri": {
						"type": "string"
					},
					"wateringDates": {
						"items": {
							"type": "string"
						},
						"type": "array"
					}
				},
				"required": [
					"id",
					"name",
					"description",
					"type",
					"sunlight",
					"wateringFrequency",
					"wateringDescription",
					"imageUri",
					"wateringDates"
				],
				"type": "object",
				"description": "From T, pick a set of properties whose keys are in the union K"
			},
			"Omit_IPlant._id_": {
				"$ref": "#/components/schemas/Pick_IPlant.Exclude_keyofIPlant._id__",
				"description": "Construct a type with the properties of T except for those in type K."
			}
		},
		"securitySchemes": {}
	},
	"info": {
		"title": "TSOA",
		"version": "1.0.0",
		"description": "Build swagger-compliant REST APIs using TypeScript and Node",
		"license": {
			"name": "MIT"
		},
		"contact": {
			"name": "unknown"
		}
	},
	"openapi": "3.0.0",
	"paths": {
		"/api/v1/plants/{plantId}": {
			"get": {
				"operationId": "GetPlant",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/IPlant"
								}
							}
						}
					},
					"401": {
						"description": "Request not authorized.",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/UnauthorizedError"
								}
							}
						}
					},
					"404": {
						"description": "Record not found.",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/NotFoundError"
								}
							}
						}
					}
				},
				"tags": [
					"Plants"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "plantId",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			},
			"put": {
				"operationId": "UpdatePlant",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/IPlant"
								}
							}
						}
					},
					"401": {
						"description": "Request not authorized.",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/UnauthorizedError"
								}
							}
						}
					},
					"404": {
						"description": "Record not found.",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/NotFoundError"
								}
							}
						}
					}
				},
				"tags": [
					"Plants"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "plantId",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/Omit_IPlant._id_"
							}
						}
					}
				}
			}
		},
		"/api/v1/plants": {
			"get": {
				"operationId": "GetAllPlants",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"items": {
										"$ref": "#/components/schemas/IPlant"
									},
									"type": "array"
								}
							}
						}
					},
					"401": {
						"description": "Request not authorized.",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/UnauthorizedError"
								}
							}
						}
					}
				},
				"tags": [
					"Plants"
				],
				"security": [],
				"parameters": []
			}
		}
	},
	"servers": [
		{
			"url": "/"
		}
	]
}